requirements.txt
pandas
numpy
requests
gspread
PyYAML